requirement.txt
groq
python-dotenv
pillow 
pytesseract
